add segment definition
