+++ SegmentDefinition.h
@@ -1,0 +1,95 @@
+#ifndef SEGMENTDEFINITION_H
+#define SEGMENTDEFINITION_H
+
+struct Enemy {
+    uint8_t type;
+    uint8_t x;
+    uint8_t y;
+    uint8_t width;
+    uint8_t height;
+    uint8_t cx;
+    uint8_t cy;
+    uint8_t cr;
+};
+
+struct Coin {
+    uint8_t show;  
+    uint8_t x;
+    uint8_t y;
+};
+
+struct SegmentDefinition {
+    uint8_t groundArray[GROUND_DEFINITION_SIZE];
+    Enemy enemies[MAX_ENEMIES_PER_SEGMENT];  // Define MAX_ENEMIES_PER_SEGMENT as needed
+    Coin coins[MAX_COINS_PER_SEGMENT];       // Define MAX_COINS_PER_SEGMENT as needed
+};
+
+const SegmentDefinition groundDefinitions[GROUND_DEFINITION_COUNT] PROGMEM = {
+    {
+      // Ground array for segment 0
+      {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1},
+      // Enemies for segment 0
+      {
+
+      },
+      {
+
+      }
+    },
+    {
+      // Ground array for segment 0
+      {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1},
+      // Enemies for segment 0
+      {
+      },
+      {
+
+      }
+    },
+    {
+      // Ground array for segment 0
+      {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1},
+      // Enemies for segment 0
+      {
+        {1, 20, 28, 32, 32, 14, 20, 10}, // Enemy 1
+      },
+      {
+
+      }
+    },
+    {
+      // Ground array for segment 0
+      {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1},
+      // Enemies for segment 0
+      {
+         {2, 20, 0, 32, 32, 12, 20, 10}, // Enemy 1
+      },
+      {
+
+      }
+    },
+    {
+      // Ground array for segment 0
+      {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1},
+      // Enemies for segment 0
+      {
+        {1, 50, 28, 32, 32, 14, 20, 10}, // Enemy 1
+      },
+      {
+
+      }
+    },
+    {
+      // Ground array for segment 0
+      {1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1},
+      // Enemies for segment 0
+      {
+      },
+      {
+
+      }
+    },
+};
+
+#endif // SEGMENTDEFINITION_H
+
